Moved localGuideProvider out of ParticleBehavior.
This guide provider now has to be set in the velocity providers
(and a few other places) where it is being used.

Related methods were removed from Particle and AbstractParticle.

Some code cleanup.
