+++ library/src/main/java/net/sourceforge/cilib/pso/particle/AbstractParticle.java
@@ -95,15 +95,6 @@ public abstract class AbstractParticle extends AbstractEntity implements Particl
     public abstract StructuredType getVelocity();
 
     /**
-     * Get the local guide of the {@code Particle}.
-     * @return A {@link StructuredType} representing the {@code Particle}'s local guide.
-     */
-    @Override
-    public StructuredType getLocalGuide() {
-        return this.behavior.getLocalGuideProvider().get(this);
-    }
-
-    /**
      * Set the neighbourhood best particle for the current Particle based on the
      * topology of the current particle.
      *
@@ -166,22 +157,6 @@ public abstract class AbstractParticle extends AbstractEntity implements Particl
     @Override
     public void setVelocityProvider(VelocityProvider velocityProvider) {
         this.behavior.setVelocityProvider(velocityProvider);
-    }
-
-    /**
-     * Get the current local {@link GuideProvider} associated with this {@code Particle}.
-     * @return The currently associated local {@link GuideProvider}.
-     */
-    public GuideProvider getLocalGuideProvider() {
-        return this.behavior.getLocalGuideProvider();
-    }
-
-    /**
-     * Set the {@link GuideProvider} for the {@code Particle}.
-     * @param localGuideProvider The local {@link GuideProvider} to use.
-     */
-    public void setLocalGuideProvider(GuideProvider localGuideProvider) {
-        this.behavior.setLocalGuideProvider(localGuideProvider);
     }
 
     /**
