+++ library/src/main/java/net/sourceforge/cilib/pso/particle/ParticleBehavior.java
@@ -25,8 +25,6 @@ public class ParticleBehavior implements Comparable<ParticleBehavior>, Cloneable
     private PositionProvider positionProvider;
     private VelocityProvider velocityProvider;
 
-    private GuideProvider localGuideProvider;
-
     private int successCounter;
     private int selectedCounter;
 
@@ -37,8 +35,6 @@ public class ParticleBehavior implements Comparable<ParticleBehavior>, Cloneable
     public ParticleBehavior() {
         this.positionProvider = new StandardPositionProvider();
         this.velocityProvider = new StandardVelocityProvider();
-
-        this.localGuideProvider = new PBestGuideProvider();
 
         this.successCounter = 0;
         this.selectedCounter = 0;
@@ -66,7 +62,6 @@ public class ParticleBehavior implements Comparable<ParticleBehavior>, Cloneable
     public ParticleBehavior(ParticleBehavior copy) {
         this.positionProvider = copy.positionProvider.getClone();
         this.velocityProvider = copy.velocityProvider.getClone();
-        this.localGuideProvider = copy.localGuideProvider.getClone();
         this.selectedCounter = copy.selectedCounter;
         this.successCounter = copy.successCounter;
     }
@@ -112,22 +107,6 @@ public class ParticleBehavior implements Comparable<ParticleBehavior>, Cloneable
      */
     public void setVelocityProvider(VelocityProvider velocityProvider) {
         this.velocityProvider = velocityProvider;
-    }
-
-    /**
-     * Get the current local <tt>GuideProvider</tt>.
-     * @return The currently associated local <tt>GuideProvider</tt>.
-     */
-    public GuideProvider getLocalGuideProvider() {
-        return this.localGuideProvider;
-    }
-
-    /**
-     * Set the <tt>GuideProvider</tt>.
-     * @param localGuideProvider The local <tt>GuideProvider</tt> to use.
-     */
-    public void setLocalGuideProvider(GuideProvider localGuideProvider) {
-        this.localGuideProvider = localGuideProvider;
     }
 
     /**
