+++ library/src/main/java/net/sourceforge/cilib/pso/velocityprovider/ConstrictionVelocityProvider.java
@@ -11,6 +11,7 @@ import net.sourceforge.cilib.controlparameter.ControlParameter;
 import net.sourceforge.cilib.math.random.generator.Rand;
 import net.sourceforge.cilib.pso.guideprovider.GuideProvider;
 import net.sourceforge.cilib.pso.guideprovider.NBestGuideProvider;
+import net.sourceforge.cilib.pso.guideprovider.PBestGuideProvider;
 import net.sourceforge.cilib.pso.particle.Particle;
 import net.sourceforge.cilib.type.types.container.Vector;
 
@@ -89,6 +90,7 @@ public class ConstrictionVelocityProvider implements VelocityProvider {
     private ControlParameter constrictionCoefficient;
     
     private GuideProvider globalGuideProvider;
+    private GuideProvider localGuideProvider;
 
     /**
      * Default constructor. The values given to the control parameters attempt to
@@ -103,6 +105,7 @@ public class ConstrictionVelocityProvider implements VelocityProvider {
         this.constrictionCoefficient = null;
         
         this.globalGuideProvider = new NBestGuideProvider();
+        this.localGuideProvider = new PBestGuideProvider();
     }
 
     /**
@@ -115,6 +118,7 @@ public class ConstrictionVelocityProvider implements VelocityProvider {
         this.kappa = copy.kappa.getClone();
         
         this.globalGuideProvider = copy.globalGuideProvider.getClone();
+        this.localGuideProvider = copy.localGuideProvider.getClone();
     }
 
     /**
@@ -138,7 +142,7 @@ public class ConstrictionVelocityProvider implements VelocityProvider {
 
         Vector velocity = (Vector) particle.getVelocity();
         Vector position = (Vector) particle.getCandidateSolution();
-        Vector localGuide = (Vector) particle.getLocalGuide();
+        Vector localGuide = (Vector) localGuideProvider.get(particle);
         Vector globalGuide = (Vector) globalGuideProvider.get(particle);
 
         Vector.Builder builder = Vector.newBuilder();
@@ -241,5 +245,13 @@ public class ConstrictionVelocityProvider implements VelocityProvider {
      */
     public void setGlobalGuideProvider(GuideProvider globalGuideProvider) {
         this.globalGuideProvider = globalGuideProvider;
+    }
+
+    /**
+     * Sets the GuideProvider responsible for retrieving a particle's local guide.
+     * @param localGuideProvider The guide provider to set.
+     */
+    public void setLocalGuideProvider(GuideProvider localGuideProvider) {
+        this.localGuideProvider = localGuideProvider;
     }
 }
