+++ library/src/main/java/net/sourceforge/cilib/pso/velocityprovider/LinearVelocityProvider.java
@@ -12,6 +12,7 @@ import net.sourceforge.cilib.controlparameter.ControlParameter;
 import net.sourceforge.cilib.math.random.generator.Rand;
 import net.sourceforge.cilib.pso.guideprovider.GuideProvider;
 import net.sourceforge.cilib.pso.guideprovider.NBestGuideProvider;
+import net.sourceforge.cilib.pso.guideprovider.PBestGuideProvider;
 import net.sourceforge.cilib.pso.particle.Particle;
 import net.sourceforge.cilib.type.types.container.Vector;
 
@@ -29,6 +30,7 @@ public class LinearVelocityProvider implements VelocityProvider {
     protected ControlParameter cognitiveAcceleration;
     
     private GuideProvider globalGuideProvider;
+    private GuideProvider localGuideProvider;
 
     /**
      * Create an instance of {@linkplain LinearVelocityProvider}.
@@ -41,6 +43,7 @@ public class LinearVelocityProvider implements VelocityProvider {
         this.cognitiveAcceleration = ConstantControlParameter.of(1.496180);
         
         this.globalGuideProvider = new NBestGuideProvider();
+        this.localGuideProvider = new PBestGuideProvider();
     }
 
     public LinearVelocityProvider(LinearVelocityProvider copy) {
@@ -48,6 +51,7 @@ public class LinearVelocityProvider implements VelocityProvider {
         this.socialAcceleration = copy.socialAcceleration.getClone();
         this.cognitiveAcceleration = copy.cognitiveAcceleration.getClone();
         this.globalGuideProvider = copy.globalGuideProvider.getClone();
+        this.localGuideProvider = copy.localGuideProvider.getClone();
     }
 
     @Override
@@ -62,7 +66,7 @@ public class LinearVelocityProvider implements VelocityProvider {
     public Vector get(Particle particle) {
         Vector velocity = (Vector) particle.getVelocity();
         Vector position = (Vector) particle.getCandidateSolution();
-        Vector localGuide = (Vector) particle.getLocalGuide();
+        Vector localGuide = (Vector) localGuideProvider.get(particle);
         Vector globalGuide = (Vector) globalGuideProvider.get(particle);
 
         float social = Rand.nextFloat();
@@ -84,5 +88,13 @@ public class LinearVelocityProvider implements VelocityProvider {
      */
     public void setGlobalGuideProvider(GuideProvider globalGuideProvider) {
         this.globalGuideProvider = globalGuideProvider;
+    }
+
+    /**
+     * Sets the GuideProvider responsible for retrieving a particle's local guide.
+     * @param localGuideProvider The guide provider to set.
+     */
+    public void setLocalGuideProvider(GuideProvider localGuideProvider) {
+        this.localGuideProvider = localGuideProvider;
     }
 }
