+++ library/src/main/java/net/sourceforge/cilib/pso/velocityprovider/MOVelocityProvider.java
@@ -11,6 +11,7 @@ import net.sourceforge.cilib.controlparameter.ConstantControlParameter;
 import net.sourceforge.cilib.controlparameter.ControlParameter;
 import net.sourceforge.cilib.pso.guideprovider.GuideProvider;
 import net.sourceforge.cilib.pso.guideprovider.NBestGuideProvider;
+import net.sourceforge.cilib.pso.guideprovider.PBestGuideProvider;
 import net.sourceforge.cilib.pso.particle.Particle;
 import net.sourceforge.cilib.math.random.generator.Rand;
 import net.sourceforge.cilib.type.types.container.Vector;
@@ -27,6 +28,7 @@ public final class MOVelocityProvider implements VelocityProvider {
     protected ControlParameter cognitiveAcceleration;
     
     private GuideProvider globalGuideProvider;
+    private GuideProvider localGuideProvider;
 
     /** Creates a new instance of StandardVelocityUpdate. */
     public MOVelocityProvider() {
@@ -41,6 +43,7 @@ public final class MOVelocityProvider implements VelocityProvider {
         this.cognitiveAcceleration = cog;
         
         this.globalGuideProvider = new NBestGuideProvider();
+        this.localGuideProvider = new PBestGuideProvider();
     }
 
     /**
@@ -52,6 +55,7 @@ public final class MOVelocityProvider implements VelocityProvider {
         this.cognitiveAcceleration = copy.cognitiveAcceleration.getClone();
         this.socialAcceleration = copy.socialAcceleration.getClone();
         this.globalGuideProvider = copy.globalGuideProvider.getClone();
+        this.localGuideProvider = copy.localGuideProvider.getClone();
     }
 
     /**
@@ -72,7 +76,7 @@ public final class MOVelocityProvider implements VelocityProvider {
         Vector velocity = (Vector) particle.getVelocity();
         Vector position = (Vector) particle.getCandidateSolution();
         Vector gbest = (Vector) particle.getNeighbourhoodBest().getCandidateSolution();
-        Vector localGuide = (Vector) particle.getLocalGuide();
+        Vector localGuide = (Vector) localGuideProvider.get(particle);
         Vector globalGuide = (Vector) globalGuideProvider.get(particle);
 
         int min = Math.min(localGuide.size(), globalGuide.size());
@@ -153,5 +157,13 @@ public final class MOVelocityProvider implements VelocityProvider {
      */
     public void setGlobalGuideProvider(GuideProvider globalGuideProvider) {
         this.globalGuideProvider = globalGuideProvider;
+    }
+
+    /**
+     * Sets the GuideProvider responsible for retrieving a particle's local guide.
+     * @param localGuideProvider The guide provider to set.
+     */
+    public void setLocalGuideProvider(GuideProvider localGuideProvider) {
+        this.localGuideProvider = localGuideProvider;
     }
 }
