+++ library/src/main/java/net/sourceforge/cilib/pso/velocityprovider/binary/BinaryQuantumVelocityProvider.java
@@ -10,6 +10,7 @@ import net.sourceforge.cilib.controlparameter.ConstantControlParameter;
 import net.sourceforge.cilib.controlparameter.ControlParameter;
 import net.sourceforge.cilib.pso.guideprovider.GuideProvider;
 import net.sourceforge.cilib.pso.guideprovider.NBestGuideProvider;
+import net.sourceforge.cilib.pso.guideprovider.PBestGuideProvider;
 import net.sourceforge.cilib.pso.particle.Particle;
 import net.sourceforge.cilib.pso.velocityprovider.VelocityProvider;
 import net.sourceforge.cilib.type.types.container.Vector;
@@ -37,6 +38,7 @@ public final class BinaryQuantumVelocityProvider implements VelocityProvider {
     protected ControlParameter cognitiveAcceleration;
     
     private GuideProvider globalGuideProvider;
+    private GuideProvider localGuideProvider;
 
     public BinaryQuantumVelocityProvider() {
         this(ConstantControlParameter.of(0.3),
@@ -60,6 +62,7 @@ public final class BinaryQuantumVelocityProvider implements VelocityProvider {
         this.cognitiveAcceleration = cognitiveAcceleration;
         
         this.globalGuideProvider = new NBestGuideProvider();
+        this.localGuideProvider = new PBestGuideProvider();
     }
 
     public BinaryQuantumVelocityProvider(BinaryQuantumVelocityProvider copy) {
@@ -69,6 +72,7 @@ public final class BinaryQuantumVelocityProvider implements VelocityProvider {
         this.socialAcceleration = copy.socialAcceleration.getClone();
         this.cognitiveAcceleration = copy.cognitiveAcceleration.getClone();
         this.globalGuideProvider = copy.globalGuideProvider.getClone();
+        this.localGuideProvider = copy.localGuideProvider.getClone();
     }
 
     /**
@@ -82,7 +86,7 @@ public final class BinaryQuantumVelocityProvider implements VelocityProvider {
     @Override
     public Vector get(Particle particle) {
         Vector velocity = (Vector) particle.getVelocity();
-        Vector pbest = (Vector) particle.getLocalGuide();
+        Vector pbest = (Vector) localGuideProvider.get(particle);
         Vector gbest = (Vector) globalGuideProvider.get(particle);
 
         Vector.Builder groupBest = Vector.newBuilder();
@@ -205,5 +209,13 @@ public final class BinaryQuantumVelocityProvider implements VelocityProvider {
      */
     public void setGlobalGuideProvider(GuideProvider globalGuideProvider) {
         this.globalGuideProvider = globalGuideProvider;
+    }
+
+    /**
+     * Sets the GuideProvider responsible for retrieving a particle's local guide.
+     * @param localGuideProvider The guide provider to set.
+     */
+    public void setLocalGuideProvider(GuideProvider localGuideProvider) {
+        this.localGuideProvider = localGuideProvider;
     }
 }
