+++ library/src/main/java/net/sourceforge/cilib/util/functions/Particles.java
@@ -33,24 +33,6 @@ public final class Particles {
         };
     }
 
-    /*public static <P extends Particle> F<P, Vector> getGlobalGuide() {
-        return new F<P, Vector>() {
-            @Override
-            public Vector f(P a) {
-                return (Vector) a.getGlobalGuide();
-            }
-        };
-    }*/
-
-    public static <P extends Particle> F<P, Vector> getLocalGuide() {
-        return new F<P, Vector>() {
-            @Override
-            public Vector f(P a) {
-                return (Vector) a.getLocalGuide();
-            }
-        };
-    }
-
     public static <P extends Particle> F<P, Particle> getNeighbourhoodBest() {
         return new F<P, Particle>() {
             @Override
